Rename package (#62)
Renames the `wikidataStuff` package to `wikidatastuff` per PEP8.

Adds both names to setup.py (temporarily for 0.4) to ensure installations
via pip of the old name remains functional while also providing the
renamed package.

Also add a a note about all of this to the README and bring it
up to date.
