+++ setup.cfg
@@ -3,4 +3,4 @@ verbosity=3
 with-coverage=1
 cover-erase=1
 cover-inclusive=1
-cover-package=wikidataStuff
+cover-package=wikidataStuff,wikidatastuff
