+++ tox.ini
@@ -14,8 +14,9 @@ deps = flake8
 commands = flake8
 
 [flake8]
-filename = 
+filename =
     wikidataStuff/*.py
+    wikidatastuff/*.py
     tests/*.py
 ignore = E501
 
@@ -24,5 +25,5 @@ deps = pydocstyle
 commands = pydocstyle
 
 [pydocstyle]
-match-dir = wikidataStuff
+match-dir = wikidata(S|s)tuff
 match = (?![test_|__init__|user\-config]).*\.py
