+++ wikidataStuff/WdqToWdqs.py
@@ -3,8 +3,8 @@
 """Backwards compatibility shim."""
 from __future__ import unicode_literals
 
-import wikidataStuff.wdq_to_wdqs
+import wikidatastuff.wdq_to_wdqs
 import wikidataStuff.deprecator
 
 wikidataStuff.deprecator.deprecate_all_functions(
-    __name__, wikidataStuff.wdq_to_wdqs)
+    __name__, wikidatastuff.wdq_to_wdqs)
